requirements.txt
Flask==2.0.2
opensearch_py==1.0.0
python_whois==0.7.3
Flask-RESTful==0.3.9
python-dotenv==0.19.2
pytest==6.2.5